Update to IPython 4.0
The new IPython splits out the parallel processing used in GEMINI into
a separate module ipyparallel, that ipython-cluster-helper now supports.
This upgrades the dependencies for ipyparallel (which pulls in ipython).
The transition can be less than smooth if pip results in partial
upgrades of a system but the APIs and imports change substantially so
it's worth making the update.