Simplify installation of external tools
Remove requirement for CloudBioLinux and Homebrew (which also needed
git) and convert to using conda package for grabix dependency, linking
into the tool directory. This is a simplification step towards moving
the full installation to conda packages so the install scripts can
get reduced to a conda installation plus symlinking and data download.

Fixes tool issues with:
https://groups.google.com/d/msg/gemini-variation/zaibTz29Te0/wL8VvB-DBwAJ