update gnomad to r2.1
this also uses a gnomad BCF which is smaller and faster to parse, and it
uses cyvcf2 (for gnomad) which is much faster than pysam.

Future version should do all VCF, BCF querying with cyvcf2, but that can
be left for another time. This commit will serve as reference for how to
do that.

see #912